Optimized code, added GameOverScreen
Cleaned some code. The game now starts in full screen mode and uses display metrics (width and height) as reference for drawing elements on screen. Added a Game Over Screen and Libgdx Prefecences to save/load previous Score and personal Bests locally on the device where the app is used.